- removed unfreeze objects - refactored and added task service test

+++ src/app/task/task-service.js
@@ -3,76 +3,70 @@
 const {Success} = require('../../common/result');
 
 module.exports = function TaskService(database) {
-    return Object.freeze({
+    return {
         createTask,
+        deleteById,
         getAllTasks,
         getTaskById,
         updateTaskById,
-        deleteById,
-    });
+    };
 
     async function createTask(task) {
-        const result = await database.insertOne(task);
+        const result  = await database.insertOne(task);
         const message = 'Task created successfully';
-        return result.withMessage(message);
+        return result.withMessage(message).setCode(201).setType('Task');
     }
 
     async function getAllTasks() {
-        return await database.findMany();
+        const result = await database.findMany();
+        return result.setCode(200).setType('Tasks');
     }
 
     async function getTaskById(taskId) {
         const result = await database.findById(taskId);
 
         return result.fold({
-            onSuccess: () => result,
+            onSuccess: () => result.setType('Task').setCode(200),
             onFailure: (failure) => {
                 const message = `Could not find a task with id ${taskId}`;
-                return failure.withMessage(message);
+                return failure.setCode(404).withMessage(message);
             }
         });
     }
 
     async function updateTaskById(taskId, task) {
-        const taskExists = await database.findById(taskId);
+        const taskExists = await getTaskById(taskId);
+        if (taskExists.isFailure()) return taskExists;
 
-        if (taskExists.isSuccess()) {
-            const query = {_id: taskId};
-            const updateResult = await database.updateOne({
-                query: query,
-                updates: task
-            });
+        const query        = {_id: taskId};
+        const updateResult = await database.updateOne({
+            query  : query,
+            updates: task
+        });
 
-            return updateResult.fold({
-                onSuccess: async () => {
-                    const message = 'Task updated successfully';
-                    const updatedResult = await getTaskById(taskId);
-                    return updatedResult.withMessage(message);
-                },
-                onFailure: (failure) => {
-                    const message = 'Unable to update a task';
-                    return failure.withMessage(message);
-                }
-            });
-        } else {
-            const message = `Could not find a task with id ${taskId}`;
-            return taskExists.withMessage(message);
-        }
+        return updateResult.fold({
+            onFailure: onTaskUpdateFailed,
+            onSuccess: onTaskUpdateSuccess,
+        });
+    }
+
+    async function onTaskUpdateSuccess(taskId) {
+        const message = 'Task updated successfully';
+        const result  = await getTaskById(taskId);
+        return result.setCode(200).withMessage(message);
+    }
+
+    async function onTaskUpdateFailed(failure) {
+        const message = 'Unable to update a task';
+        return failure.setCode(400).withMessage(message);
     }
 
     async function deleteById(taskId) {
-        const result = await database.findById(taskId);
+        const result = await getTaskById(taskId);
+        if (result.isFailure()) return result;
 
-        return result.fold({
-            onSuccess: async () => {
-                await database.deleteOne({_id: taskId});
-                const message = 'Task deleted successfully.';
-                return new Success(message);
-            },
-            onFailure: (failure) => {
-                const message = `Could not find a task with id ${taskId}`;
-                return failure.withMessage(message);
-            }
-        });
+        await database.deleteOne({_id: taskId});
+        const message = 'Task deleted successfully.';
+        return new Success(message).setCode(200);
     }
 };
